+++ ui/login/src/main/java/dev/mkeeda/spaceship/ui/login/password/presentation/PasswordLoginEffect.kt
@@ -1,4 +1,4 @@
-package dev.mkeeda.spaceship.ui.login.presentation
+package dev.mkeeda.spaceship.ui.login.password.presentation
 
 sealed class PasswordLoginEffect {
     object NavigateToMain : PasswordLoginEffect()
