+++ ui/login/src/main/java/dev/mkeeda/spaceship/ui/login/password/presentation/PasswordLoginViewModel.kt
@@ -1,4 +1,4 @@
-package dev.mkeeda.spaceship.ui.login.presentation
+package dev.mkeeda.spaceship.ui.login.password.presentation
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
